scalar/scalar-azrepos: add watchman dependency
Add a dependency on the `watchman` formula for both the `scalar` and
`scalar-azrepos` Casks. Although Watchman is not a hard requirement to
use Scalar, it is recommended to get the best performance (which is
probably the reason for using Scalar in the first place).

Once the built-in FSMonitor is working on all platforms, we can drop
this dependency.